Renamed getFiles to ixMaterials
