+++ app/api/suggestions/blankSuggestions.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-statements */
 import { EnforcedWithId } from 'api/odm';
 import { DefaultDispatcher } from 'api/queue.v2/configuration/factories';
-import { propertyTypeIsMultiValued } from 'api/services/informationextraction/getFiles';
+import { propertyTypeIsMultiValued } from 'api/services/informationextraction/ixMaterials';
 import templates from 'api/templates';
 import { tenants } from 'api/tenants';
 import { LanguageUtils } from 'shared/language';
